# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "operationId": "submitChat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Chat"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "UI Message Stream v1 SSE; consume with AI SDK",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/{id}/stream": {
      "get": {
        "tags": [
          "chat"
        ],
        "operationId": "reconnectChat",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "UI Message Stream v1 replay",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversations": {
      "get": {
        "tags": [
          "conversations"
        ],
        "operationId": "listConversations",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConversationView"
                  }
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "conversations"
        ],
        "operationId": "createConversation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Create"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationView"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversations/{id}": {
      "get": {
        "tags": [
          "conversations"
        ],
        "operationId": "getConversation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationView"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversations/{id}/commands/{request_id}": {
      "get": {
        "tags": [
          "conversations"
        ],
        "operationId": "getCommandResult",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "operationId": "getRun",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{id}/{action}": {
      "post": {
        "tags": [
          "runs"
        ],
        "operationId": "controlRun",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RunAction"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "operationId": "getHealth",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "AttemptView": {
        "type": "object",
        "required": [
          "id",
          "runId",
          "generation",
          "modelCalls",
          "toolCalls",
          "knownTokens",
          "usageComplete"
        ],
        "properties": {
          "generation": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string"
          },
          "knownTokens": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "modelCalls": {
            "type": "integer",
            "minimum": 0
          },
          "outcome": {
            "type": [
              "string",
              "null"
            ]
          },
          "runId": {
            "type": "string"
          },
          "toolCalls": {
            "type": "integer",
            "minimum": 0
          },
          "traceId": {
            "type": [
              "string",
              "null"
            ]
          },
          "usageComplete": {
            "type": "boolean"
          }
        }
      },
      "Chat": {
        "type": "object",
        "required": [
          "id",
          "message",
          "expectedRevision",
          "requestId"
        ],
        "properties": {
          "expectedRevision": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string"
          },
          "message": {
            "$ref": "#/components/schemas/UiMessageView"
          },
          "requestId": {
            "type": "string"
          },
          "trigger": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false
      },
      "Command": {
        "type": "object",
        "required": [
          "expectedVersion",
          "requestId",
          "conversationId"
        ],
        "properties": {
          "conversationId": {
            "type": "string"
          },
          "expectedRevision": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "expectedVersion": {
            "type": "integer",
            "format": "int64"
          },
          "message": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/UiMessageView"
              },
              {
                "type": "null"
              }
            ]
          },
          "requestId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "CommandResult": {
        "type": "object",
        "required": [
          "runId"
        ],
        "properties": {
          "runId": {
            "type": "string"
          }
        }
      },
      "ConversationView": {
        "type": "object",
        "required": [
          "id",
          "revision",
          "messages"
        ],
        "properties": {
          "activeRunId": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UiMessageView"
            }
          },
          "parentId": {
            "type": [
              "string",
              "null"
            ]
          },
          "revision": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "Create": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UiMessageView"
            }
          }
        },
        "additionalProperties": false
      },
      "Health": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
      "MessageRole": {
        "type": "string",
        "enum": [
          "system",
          "user",
          "assistant"
        ]
      },
      "RunAction": {
        "type": "string",
        "enum": [
          "pause",
          "resume",
          "cancel",
          "steer"
        ]
      },
      "RunDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RunView"
          },
          {
            "type": "object",
            "required": [
              "statistics",
              "attempts"
            ],
            "properties": {
              "attempts": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AttemptView"
                }
              },
              "statistics": {
                "$ref": "#/components/schemas/RunStatistics"
              }
            }
          }
        ]
      },
      "RunStatistics": {
        "type": "object",
        "required": [
          "modelCalls",
          "toolCalls",
          "knownTokens",
          "usageComplete"
        ],
        "properties": {
          "knownTokens": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "modelCalls": {
            "type": "integer",
            "minimum": 0
          },
          "toolCalls": {
            "type": "integer",
            "minimum": 0
          },
          "usageComplete": {
            "type": "boolean"
          }
        }
      },
      "RunStatusView": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "pausing",
          "paused",
          "needs-attention",
          "finished",
          "step-limit",
          "failed",
          "cancelled",
          "superseded"
        ]
      },
      "RunView": {
        "type": "object",
        "required": [
          "id",
          "conversationId",
          "assistantId",
          "status",
          "version",
          "generation",
          "steps"
        ],
        "properties": {
          "assistantId": {
            "type": "string"
          },
          "attemptId": {
            "type": [
              "string",
              "null"
            ]
          },
          "conversationId": {
            "type": "string"
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "generation": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatusView"
          },
          "steps": {
            "type": "integer",
            "minimum": 0
          },
          "supersedes": {
            "type": [
              "string",
              "null"
            ]
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "UiMessageView": {
        "type": "object",
        "description": "AI SDK UIMessage envelope. Parts follow UI Message Stream v1.",
        "required": [
          "id",
          "role",
          "parts"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "metadata": {},
          "parts": {
            "type": "array",
            "items": {}
          },
          "role": {
            "$ref": "#/components/schemas/MessageRole"
          }
        }
      }
    }
  }
}